Refactor Maze to static only
Since the old generate() function could only be called once, it makes
more sense for the function and the functions it calls to be static,
since the class doesn't need any instance variables.

The new run(options, canvas) function will return a list of
AnimationTimers, which the recipient can control themselves (stop it).

Because the run function handles the timers, the timer classes are no
longer needed.